requirements.txt
foursquare==1!2016.9.12
googlemaps==2.4.6
langid==1.1.6
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
pyquery==1.2.17
requests==2.13.0
Scrapy==1.3.3
Shapely==1.5.17.post1
wget==3.2
beautifulsoup4==4.5.3
PyGithub==1.34
pyshp==1.2.10
python_twitter==3.2.1
